requirements.txt
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.0
pandas==2.0.3
functions-framework==3.4.0